Simple workspace structure
