+++ Cargo.toml
@@ -1,6 +1,8 @@
-[package]
-name = "arena"
-version = "0.1.0"
-edition = "2024"
+[workspace]
+members = [
+    "arena",
+    "examples"
+]
 
-[dependencies]
+[workspace.dependencies]
+arena = { path = "arena" }
