+++ arena/src/arena.rs
@@ -1,21 +1,21 @@
 use std::ops::Drop;
-use super::a_match::AMatch;
+use super::arena_match::ArenaMatch;
 
 pub struct Arena {
     pub name: String,
-    pub arena_matches: Vec<AMatch>,
+    pub encounters: Vec<ArenaMatch>,
     running: bool
 }
 
 impl Arena {
-    pub fn new(name: String, arena_matches: Vec<AMatch>) -> Self {
-        Arena { name, arena_matches, running: false }
+    pub fn new(name: String, encounters: Vec<ArenaMatch>) -> Self {
+        Arena { name, encounters, running: false }
     }
 
     pub fn commence(&mut self) {
         if self.running { return; }
         println!("[ARENA:{}] Starting.", self.name);
-        for m in self.arena_matches.iter_mut() {
+        for m in self.encounters.iter_mut() {
             m.start();
         }
         self.running = true;
@@ -25,7 +25,7 @@ impl Arena {
     pub fn conclude(&mut self) {
         if !self.running { return;}
         println!("[ARENA:{}] Stopping matches.", self.name);
-        for m in self.arena_matches.iter_mut().rev() {
+        for m in self.encounters.iter_mut().rev() {
             m.stop();
         }
         println!("[ARENA:{}] Stopped matches.", self.name);
