+++ arena/src/component.rs
