+++ arena/src/lib.rs
@@ -1,0 +1,13 @@
+pub mod arena;
+pub mod dependency;
+pub mod component;
+pub mod arena_match;
+pub mod postgres_dependency;
+pub mod couchbase_dependency;
+
+pub use crate::arena::Arena;
+pub use crate::arena_match::ArenaMatch;
+pub use crate::component::{Component, ManagedProcessComponent};
+pub use crate::dependency::Dependency;
+pub use crate::postgres_dependency::PostgresDependency;
+pub use crate::couchbase_dependency::CouchbaseDependency;
