+++ arena/src/postgres_dependency.rs
@@ -1,35 +1,34 @@
-use crate::arena::dependency::RunnableDependency;
-use crate::Dependency;
+use crate::dependency::{Dependency, RunnableDependency};
 
-pub struct CouchbaseDependency {
+pub struct PostgresDependency {
     pub name: String,
     pub dependencies: Vec<Dependency>,
     started: bool
 }
 
-impl CouchbaseDependency {
+impl PostgresDependency {
     pub fn new(name: String) -> Self {
-        CouchbaseDependency { name, dependencies: vec![], started: false }
+        PostgresDependency { name, dependencies: vec![], started: false }
     }
 }
 
-impl RunnableDependency for CouchbaseDependency {
+impl RunnableDependency for PostgresDependency {
     fn start(&mut self) {
         if self.started { return; }
-        println!("[{}] (Couchbase) Starting connection.", self.name);
+        println!("[{}] (DB) Starting connection.", self.name);
         for dep in self.dependencies.iter_mut() {
             dep.start();
         }
         self.started = true;
-        println!("[{}] (Couchbase) Connection started.", self.name);
+        println!("[{}] (DB) Connection started.", self.name);
     }
 
     fn stop(&mut self) {
-        println!("[{}] (Couchbase) Stopping connection.", self.name);
+        println!("[{}] Stopping connection.", self.name);
         for dep in self.dependencies.iter_mut().rev() {
             dep.stop();
         }
-        println!("[{}] (Couchbase) Connection stopped.", self.name);
+        println!("[{}] (DB) Connection stopped.", self.name);
     }
 
     fn add_child_internal(&mut self, dep: Dependency) {
@@ -37,7 +36,7 @@ impl RunnableDependency for CouchbaseDependency {
     }
 }
 
-impl Drop for CouchbaseDependency {
+impl Drop for PostgresDependency {
     fn drop(&mut self) {
         self.stop();
     }
