--- src/arena.rs
@@ -1,6 +1,0 @@
-pub mod dependency;
-pub mod component;
-pub mod a_match;
-pub mod arena;
-pub mod postgres_dependency;
-pub mod couchbase_dependency;
