--- src/lib.rs
@@ -1,8 +1,0 @@
-pub mod arena;
-
-pub use arena::arena::Arena;
-pub use arena::a_match::AMatch;
-pub use arena::dependency::Dependency;
-pub use arena::postgres_dependency::PostgresDependency;
-pub use arena::couchbase_dependency::CouchbaseDependency;
-pub use arena::component::{Component, ManagedProcessComponent};
